refactor(budget): onClick 수정
