+++ src/pages/Budget/Budget.jsx
@@ -1,3 +1,8 @@
+import { useState, useEffect, useRef } from 'react';
+import { useDispatch, useSelector } from 'react-redux';
+import axios from 'axios';
+import Nav from '../../components/Nav';
+import Footer from '../../components/Footer';
 import {
   Container,
   Row,
@@ -7,18 +12,12 @@ import {
   Form,
   Modal,
 } from 'react-bootstrap';
-import 'bootstrap/dist/css/bootstrap.min.css';
-import Nav from '../../components/Nav';
-import Footer from '../../components/Footer';
-import { useState, useEffect, useRef } from 'react';
-import axios from 'axios';
-
-import { useDispatch, useSelector } from 'react-redux';
+import { FaArrowAltCircleUp, FaPencilAlt, FaTrash } from 'react-icons/fa';
 import { chargeUpdate } from '../../store/modules/budget';
-import BudgetReceipt from '../Budget/BudgetReceipt';
 
 export default function Budget() {
   const dispatch = useDispatch();
+  const nickName = useSelector((state) => state.users.userNickName);
 
   const textRef = useRef();
   const chargeRef = useRef();
@@ -29,9 +28,21 @@ export default function Budget() {
   const [update, setUpdate] = useState(false);
   const [okay, setOkay] = useState(false);
   const [show, setShow] = useState(false);
+  const [users, setUsers] = useState(1);
+
   const handleClose = () => setShow(false);
 
-  const nickName = useSelector((state) => state.users.userNickName);
+  let totalCharge = [];
+
+  if (chargeData !== undefined) {
+    totalCharge = chargeData?.reduce((acc, cur, i) => {
+      return cur.charge + acc;
+    }, 0);
+  }
+
+  const addUserNum = () => {
+    setUsers(users + 1);
+  };
 
   useEffect(() => {
     axios
@@ -43,6 +54,45 @@ export default function Budget() {
       })
       .catch((err) => console.log(err));
   }, [update, chargeData]);
+
+  const resetBudget = () => {
+    axios
+      .post(`http://13.125.234.1:4000/charge/alldelete`, {
+        nickName,
+        chargeData,
+      })
+      .then((결과) => {
+        console.log('초기화 성공');
+        setShow(false);
+        setUpdate(!update);
+      })
+      .catch((err) => {
+        console.log(err);
+      });
+  };
+
+  const addBudget = () => {
+    const date = dateRef.current.value;
+    const title = textRef.current.value;
+    const charge = chargeRef.current.value;
+    axios
+      .post(`http://13.125.234.1:4000/charge/write`, {
+        chargeList: { date, title, charge: parseInt(charge) },
+        nickName,
+      })
+      .then((res) => {
+        console.log('charge 등록 성공');
+        alert('여행 지출 내역 등록을 성공하였습니다🙌');
+        dateRef.current.value = '';
+        textRef.current.value = '';
+        chargeRef.current.value = '';
+        setUpdate(!update);
+      })
+      .catch(() => {
+        console.log('charge 등록 실패');
+        alert('여행 지출 내역 등록을 실패하였습니다. 다시 시도해주세요.');
+      });
+  };
 
   if (show) {
     return (
@@ -68,24 +118,7 @@ export default function Budget() {
           </p>
         </Modal.Body>
         <Modal.Footer>
-          <Button
-            variant="outline-success"
-            onClick={() => {
-              axios
-                .post(`http://13.125.234.1:4000/charge/alldelete`, {
-                  nickName,
-                  chargeData,
-                })
-                .then((결과) => {
-                  console.log('초기화 성공');
-                  setShow(false);
-                  setUpdate(!update);
-                })
-                .catch(() => {
-                  console.log('실패');
-                });
-            }}
-          >
+          <Button variant="outline-success" onClick={resetBudget}>
             초기화
           </Button>
           <Button variant="success" onClick={handleClose}>
@@ -147,30 +180,7 @@ export default function Budget() {
                   <Button
                     variant="success"
                     className="text-end"
-                    onClick={() => {
-                      const date = dateRef.current.value;
-                      const title = textRef.current.value;
-                      const charge = chargeRef.current.value;
-                      axios
-                        .post(`http://13.125.234.1:4000/charge/write`, {
-                          chargeList: { date, title, charge: parseInt(charge) },
-                          nickName,
-                        })
-                        .then((res) => {
-                          console.log('charge 등록 성공');
-                          alert('여행 지출 내역 등록을 성공하였습니다🙌');
-                          dateRef.current.value = '';
-                          textRef.current.value = '';
-                          chargeRef.current.value = '';
-                          setUpdate(!update);
-                        })
-                        .catch(() => {
-                          console.log('charge 등록 실패');
-                          alert(
-                            '여행 지출 내역 등록을 실패하였습니다. 다시 시도해주세요.'
-                          );
-                        });
-                    }}
+                    onClick={addBudget}
                   >
                     등록
                   </Button>
@@ -179,8 +189,106 @@ export default function Budget() {
             </Col>
 
             {/* 오른쪽 영수증 */}
+            <Col
+              className=" p-5 rounded border mt-4"
+              style={{ backgroundColor: '#fafafa' }}
+            >
+              <h6 className="fw-bold text-center" style={{ color: '#198754' }}>
+                TripLog
+              </h6>
+              <h2 className="fw-bold text-center">RECEIPT</h2>
 
-            <BudgetReceipt props={chargeData} />
+              <hr class="solid" style={{ borderTopWidth: '2px' }}></hr>
+
+              <Row className=" mb-2 mx-1">
+                <Col className="fw-bold col-3">Day</Col>
+                <Col className="fw-bold col-5 text-center">ITEM</Col>
+                <Col className="fw-bold col-2 text-center ">Price</Col>
+                <Col className="fw-bold col-2 text-end">Del</Col>
+              </Row>
+              <hr class="solid"></hr>
+
+              {chargeData &&
+                chargeData.map(function (a, i) {
+                  return (
+                    <Row className="mx-1" key={i}>
+                      <Col className="col-3">
+                        <p>{a.date.slice(5, 10)}</p>
+                      </Col>
+                      <Col className="col-5 text-center">{a.title}</Col>
+                      <Col className="col-2 text-center">
+                        {a.charge.toLocaleString('ko-KR', {
+                          currency: 'KRW',
+                        })}
+                      </Col>
+                      <Col
+                        className="col-2 text-end"
+                        style={{ cursor: 'pointer' }}
+                      >
+                        <FaTrash
+                          style={{ color: 'grey' }}
+                          onClick={() => {
+                            axios
+                              .post('http://13.125.234.1:4000/charge/delete', {
+                                nickName,
+                                a,
+                              })
+                              .then((결과) => {
+                                alert('지출 내역 삭제를 성공하였습니다🙌');
+                                setUpdate(!update);
+                              })
+                              .catch((err) => {
+                                console.log(err);
+                              });
+                          }}
+                        />
+                      </Col>
+                    </Row>
+                  );
+                })}
+
+              <hr class="dashed" style={{ borderTop: 'dashed' }}></hr>
+              <Row>
+                <Col sm md lg="auto" className="fw-bold">
+                  ITEM COUNT :
+                </Col>
+                <Col className="text-end">{chargeData.length} 개</Col>
+              </Row>
+
+              <Row>
+                <Col className="fw-bold">
+                  정산 : {users} 명 {'\u00A0'}
+                  <FaArrowAltCircleUp
+                    onClick={addUserNum}
+                    style={{ cursor: 'pointer', color: '#198754' }}
+                  />
+                </Col>
+                <Col sm md lg="auto" className="text-end">
+                  1인당{' '}
+                  {parseInt(totalCharge / users).toLocaleString('ko-KR', {
+                    currency: 'KRW',
+                  })}
+                  원
+                </Col>
+              </Row>
+
+              <Row>
+                <Col className="fw-bold">총 합계 : </Col>
+                <Col sm md lg="auto" className="text-end">
+                  {totalCharge.toLocaleString('ko-KR', {
+                    currency: 'KRW',
+                  })}
+                  원
+                </Col>
+              </Row>
+
+              <hr class="dashed" style={{ borderTop: 'dashed' }}></hr>
+              <Col className="text-end">
+                <Button variant="success" onClick={() => setShow(true)}>
+                  초기화
+                </Button>
+              </Col>
+            </Col>
           </Row>
         </Container>
         <Footer />
